CBD-2372 Enabling escrow build
Reviewed-on: http://review.couchbase.org/106382
Tested-by: Kim Trang <kim.trang@couchbase.com>

+++ cbdeps/protoc-gen-go/protoc-gen-go_unix.sh
@@ -4,16 +4,21 @@ INSTALL_DIR=$1
 
 DEPS=${WORKSPACE}/deps
 GO_VERS=1.11.5
+CBDEP_TOOL_VERS=0.9.3
 
-if [ $(uname -s) = "Darwin" ]; then
-    CBDEP_URL=https://packages.couchbase.com/cbdep/0.9.3/cbdep-0.9.3-darwin
+CBDEP_BIN_PATH=/home/couchbase/.cbdepscache/cbdep/${CBDEP_TOOL_VERS}/cbdep-${CBDEP_TOOL_VERS}-linux
+if [[ ! -f ${CBDEP_BIN_PATH} ]]; then
+    if [ $(uname -s) = "Darwin" ]; then
+        CBDEP_URL=https://packages.couchbase.com/cbdep/${CBDEP_TOOL_VERS}/cbdep-${CBDEP_TOOL_VERS}-darwin
+    else
+        CBDEP_URL=https://packages.couchbase.com/cbdep/${CBDEP_TOOL_VERS}/cbdep-${CBDEP_TOOL_VERS}-linux
+    fi
+    curl -o /tmp/cbdep ${CBDEP_URL}
+    chmod +x /tmp/cbdep
+    /tmp/cbdep install -d "${DEPS}" golang ${GO_VERS}
 else
-    CBDEP_URL=https://packages.couchbase.com/cbdep/0.9.3/cbdep-0.9.3-linux
+   ${CBDEP_BIN_PATH} install -d "${DEPS}" golang ${GO_VERS}
 fi
-
-curl -o /tmp/cbdep ${CBDEP_URL}
-chmod +x /tmp/cbdep
-/tmp/cbdep install -d "${DEPS}" golang ${GO_VERS}
 
 GOPATH=${WORKSPACE}
 PATH=${DEPS}/go1.11.5/bin:$PATH
